+++ pyroomacoustics/libroom_src/room.cpp
@@ -27,14 +27,8 @@
 #include <iostream>
 #include <cmath>
 #include <algorithm>
+#include "constants.hpp"
 #include "room.hpp"
-
-const double pi = 3.14159265358979323846;
-const double pi_2 = 1.57079632679489661923;
-
-// Initial energy of a particule.
-// The value 2.0 is necessary to match the scale of the ISM.
-const double energy_0_numerator = 2.0f;
 
 size_t number_image_sources_2(size_t max_order) {
   /*
@@ -956,7 +950,6 @@ void Room<D>::simul_ray(
           double r_sq = double(travel_dist_at_mic) * travel_dist_at_mic;
           auto p_hit = (1 - sqrt(1 - mic_radius_sq / std::max(mic_radius_sq, r_sq)));
           energy = transmitted / (r_sq * p_hit);
-          // energy = transmitted / (travel_dist_at_mic - sqrtf(fmaxf(0.f, travel_dist_at_mic * travel_dist_at_mic - mic_radius_sq)));
 
           microphones[k].log_histogram(travel_dist_at_mic, energy, start);
         }
@@ -968,7 +961,7 @@ void Room<D>::simul_ray(
     transmitted *= wall.get_energy_reflection();
 
     // Let's shoot the scattered ray induced by the rebound on the wall
-    if (wall.scatter.maxCoeff() > 0.f)
+    if (wall.does_scatter)
     {
       // Shoot the scattered ray
       scat_ray(
@@ -978,6 +971,7 @@ void Room<D>::simul_ray(
           hit_point,
           travel_dist
           );
+
     }
 
     // Check if we reach the thresholds for this ray
@@ -986,7 +980,18 @@ void Room<D>::simul_ray(
 
     // set up for next iteration
     specular_counter += 1;
+
     dir = wall.normal_reflect(dir);  // conserves length
+
+    if (wall.does_scatter)
+    {
+      // Compute the scattered direction.
+      // We inverse the direction because the wall normal is pointing outward.
+      Vectorf<D> scat_dir = -wall.sample_lambertian_reflection();
+      dir = wall.average_scatter * scat_dir + (1.f - wall.average_scatter) * dir;
+      dir = dir.normalized();
+    }
+
     start = hit_point;
   }
 }
@@ -999,12 +1004,12 @@ void Room<D>::ray_tracing(
   )
 {
   // float energy_0 = 2.f / (mic_radius * mic_radius * angles.cols());
-  float energy_0 = energy_0_numerator / angles.cols();
+  float energy_0 = constants::ENERGY_0 / angles.cols();
 
   for (int k(0) ; k < angles.cols() ; k++)
   {
     float phi = angles.coeff(0,k);
-    float theta = pi_2;
+    float theta = constants::HALF_PI;
 
     if (D == 3)
       theta = angles.coeff(1,k);
@@ -1039,13 +1044,13 @@ void Room<D>::ray_tracing(
 
   // ------------------ INIT --------------------
   // initial energy of one ray
-  float energy_0 = energy_0_numerator / (nb_phis * nb_thetas);
+  float energy_0 = constants::ENERGY_0 / (nb_phis * nb_thetas);
 
   // ------------------ RAY TRACING --------------------
 
   for (size_t i(0); i < nb_phis; ++i)
   {
-    float phi = 2 * pi * (float) i / nb_phis;
+    float phi = constants::TWO_PI * (float) i / nb_phis;
 
     for (size_t j(0); j < nb_thetas; ++j)
     {
@@ -1055,7 +1060,7 @@ void Room<D>::ray_tracing(
       // For 2D, this parameter means nothing, but we set it to
       // PI/2 to be consistent
       if (D == 2) {
-        theta = pi_2;
+        theta = constants::HALF_PI;
       }
 
       // Trace this ray
@@ -1089,13 +1094,13 @@ void Room<D>::ray_tracing(
 
   // ------------------ INIT --------------------
   // initial energy of one ray
-  float energy_0 = energy_0_numerator / n_rays;
+  float energy_0 = constants::ENERGY_0 / n_rays;
 
   // ------------------ RAY TRACING --------------------
   if (D == 3)
   {
     auto offset = 2.f / n_rays;
-    auto increment = pi * (3.f - sqrt(5.f));  // phi increment
+    auto increment = constants::PI * (3.f - sqrt(5.f));  // phi increment
 
     for (size_t i(0); i < n_rays ; ++i)
     {
@@ -1115,7 +1120,7 @@ void Room<D>::ray_tracing(
   }
   else if (D == 2)
   {
-    float offset = 2. * pi / n_rays;
+    float offset = constants::TWO_PI / n_rays;
     for (size_t i(0) ; i < n_rays ; ++i)
       simul_ray(i * offset, 0.f, source_pos, energy_0);
   }
@@ -1178,12 +1183,12 @@ bool Room<D>::contains(const Vectorf<D> point)
     n_intersections = 0;
     ambiguous_intersection = false;
 
-    outside_point[0] -= (float)(rand() % 27) / 50;
-    outside_point[1] -= (float)(rand() % 22) / 26;
+    outside_point[0] -= rng::uniform(0.05f, 0.54f);
+    outside_point[1] -= rng::uniform(0.05f, 0.8461f);
 
     if (D == 3)
     {
-      outside_point[2] -= (float)(rand() % 24 / 47);
+      outside_point[2] -= rng::uniform(0.05f, 0.501f);
     }
 
     for (size_t i(0); i < n_walls; ++i)