+++ pyroomacoustics/libroom_src/wall.hpp
@@ -56,6 +56,8 @@ class Wall
     // Wall properties container
     Eigen::ArrayXf absorption;  // the wall absorption coefficient for every freq. band
     Eigen::ArrayXf scatter;  // the wall scattering coefficient for every freq. band
+    float average_scatter;  // the average scattering coefficient accross all subbands
+    bool does_scatter = false;  // flag that indicates that the wall has non-zero scattering
     std::string name;
     Eigen::ArrayXf transmission;  // computed from absorption as sqrt(1 - a)
     Eigen::ArrayXf energy_reflection;  // computed from absorption as (1 - a)
@@ -84,9 +86,9 @@ class Wall
 
     // Copy constructor
     Wall(const Wall<D> &w) :
-      absorption(w.absorption), scatter(w.scatter), name(w.name),
-      transmission(w.transmission), energy_reflection(w.energy_reflection),
-      normal(w.normal), corners(w.corners),
+      absorption(w.absorption), scatter(w.scatter), average_scatter(w.average_scatter),
+      does_scatter(w.does_scatter), name(w.name), transmission(w.transmission),
+      energy_reflection(w.energy_reflection), normal(w.normal), corners(w.corners),
       origin(w.origin), basis(w.basis), flat_corners(w.flat_corners)
     {}
 
@@ -123,6 +125,8 @@ class Wall
     float cosine_angle(   // cosine angle with respect to surface normal
         const Vectorf<D> &p
         ) const;
+
+    Vectorf<D> sample_lambertian_reflection() const;
 }; 
 
 #include "wall.cpp"